+++ shared/ui/src/commonMain/kotlin/com/shub39/rush/shared/ui/lyrics/section/LyricsCustomisationPage.kt
@@ -389,7 +389,7 @@ private fun AudioPermissionDialog(
     onLaunchPermission: () -> Unit = {},
 ) {
     if (show) {
-        RushDialog(onDismissRequest = onDismiss) {
+        RushDialog(onDismissRequest = onDismiss, padding = 32.dp) {
             Column(
                 modifier = Modifier.wrapContentSize(),
                 horizontalAlignment = Alignment.CenterHorizontally,